requirements.txt
tensorflow==2.15.0
seaborn==0.12.2
matplotlib==3.8.0
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.2.2
pytest==7.4.0
python-dotenv==1.0.1
pydantic==1.10.12
